requirements.txt
pandas
#openai
python-dotenv
groq
flask